# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "zoo",
    "description": "### An asynchronous zoo API, powered by FastAPI and SQLModel\n\n[<img src=\"https://raw.githubusercontent.com/juftin/juftin/main/static/juftin.png\" width=\"120\" height=\"120\"  alt=\"juftin logo\">](https://juftin.com)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/exhibits": {
      "get": {
        "tags": [
          "exhibits"
        ],
        "summary": "Get Exhibits",
        "description": "Get exhibits from the database",
        "operationId": "exhibits-get_exhibits",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExhibitsRead"
                  },
                  "title": "Response Exhibits-Get Exhibits"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exhibits"
        ],
        "summary": "Create Exhibit",
        "description": "Create a new exhibit in the database",
        "operationId": "exhibits-create_exhibit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExhibitsCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exhibits/{exhibit_id}": {
      "get": {
        "tags": [
          "exhibits"
        ],
        "summary": "Get Exhibit",
        "description": "Get exhibit from the database",
        "operationId": "exhibits-get_exhibit",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exhibit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exhibits"
        ],
        "summary": "Delete Exhibit",
        "description": "Delete exhibit from the database",
        "operationId": "exhibits-delete_exhibit",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exhibit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "exhibits"
        ],
        "summary": "Update Exhibit",
        "description": "Update exhibit from the database",
        "operationId": "exhibits-update_exhibit",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exhibit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExhibitsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exhibits/{exhibit_id}/animals": {
      "get": {
        "tags": [
          "exhibits"
        ],
        "summary": "Get Exhibit Animals",
        "description": "List animals in an exhibit",
        "operationId": "exhibits-get_exhibit_animals",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exhibit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalsRead"
                  },
                  "title": "Response Exhibits-Get Exhibit Animals"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exhibits/{exhibit_id}/staff": {
      "get": {
        "tags": [
          "exhibits"
        ],
        "summary": "Get Exhibit Staff",
        "description": "List staff in an exhibit",
        "operationId": "exhibits-get_exhibit_staff",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Exhibit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffRead"
                  },
                  "title": "Response Exhibits-Get Exhibit Staff"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnimalsRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The unique identifier for the table"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the animal"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The description of the animal"
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species",
            "description": "The species of the animal"
          },
          "exhibit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exhibit Id",
            "description": "The id of the exhibit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The date and time the record was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The date and time the record was last modified"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "The date and time the record was deleted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "AnimalsRead",
        "description": "\n    Animals model: read\n    ",
        "examples": [
          {
            "created_at": "2021-01-01T00:00:00.000000",
            "description": "Ferocious kitty",
            "exhibit_id": 1,
            "id": 1,
            "modified_at": "2021-01-02T09:12:34.567890",
            "name": "Lion",
            "species": "Panthera leo"
          }
        ]
      },
...
      "ExhibitsCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the exhibit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The description of the exhibit"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "The location of the exhibit"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ExhibitsCreate",
        "description": "\n    Exhibits model: create\n    ",
        "examples": [
          {
            "description": "A big cat exhibit",
            "location": "North America",
            "name": "Big Cat Exhibit"
          }
        ]
      },
      "ExhibitsRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The unique identifier for the table"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the exhibit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The description of the exhibit"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "The location of the exhibit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The date and time the record was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The date and time the record was last modified"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "The date and time the record was deleted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "ExhibitsRead",
        "description": "\n    Exhibits model: read\n    ",
        "examples": [
          {
            "created_at": "2021-01-01T00:00:00.000000",
            "description": "A big cat exhibit",
            "id": 1,
            "location": "North America",
            "modified_at": "2021-01-02T09:12:34.567890",
            "name": "Big Cat Exhibit"
          }
        ]
      },
      "ExhibitsUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the exhibit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The description of the exhibit"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "The location of the exhibit"
          }
        },
        "type": "object",
        "title": "ExhibitsUpdate",
        "description": "\n    Exhibits model: update\n    ",
        "examples": [
          {
            "name": "Big Cat Exhibit"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StaffRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The unique identifier for the table"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the staff"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title",
            "description": "The job title of the staff"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "The email of the staff"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "The phone number of the staff"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional notes regarding the staff member"
          },
          "exhibit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exhibit Id",
            "description": "The id of the exhibit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The date and time the record was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The date and time the record was last modified"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "The date and time the record was deleted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "StaffRead",
        "description": "\n    Staff model: read\n    ",
        "examples": [
          {
            "created_at": "2021-01-01T00:00:00.000000",
            "email": "big.cat.lover@gmail.com",
            "exhibit_id": 1,
            "id": 1,
            "job_title": "Zookeeper",
            "modified_at": "2021-01-02T09:12:34.567890",
            "name": "John Doe",
            "notes": "John Doe is a great zookeeper and loves cats!",
            "phone": "555-555-5555"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}